+++ user_accounts/mb_side.c
@@ -9,7 +9,7 @@
 
 #define KEY ((const unsigned char *) "B&/n^!v8G`3BJL:B~q`~K(y!~;SBDw0\0")
 
-void encrypt(user_account *input_user, uint16_t size, user_account *ret_user) {
+void encrypt(user_account *input_user, uint32_t size, user_account *ret_user) {
 	for(uint16_t i=0; i < size - 1; i++) {
 		ret_user->pword[i] = input_user->pword[i]^KEY[i];
 		ret_user->uname[i] = input_user->uname[i];
@@ -18,19 +18,19 @@ void encrypt(user_account *input_user, uint16_t size, user_account *ret_user) {
 	memset(input_user->pword, 0, size);
 }
 
-void decrypt(unsigned char* cipher_p, uint16_t size, unsigned char* decrypted) {
-	for(uint16_t i=0; i < size; i++) {
+void decrypt(unsigned char* cipher_p, uint32_t size, unsigned char* decrypted) {
+	for(uint32_t i=0; i < size; i++) {
 		decrypted[i] = cipher_p[i]^KEY[i];
 	}
 }
 
-void create_user(user_account *user_data, uint16_t *pword_len, 
+void create_user(user_account *user_data, uint32_t *pword_len, 
 	user_account *user_cipher) {
 	encrypt(user_data, *pword_len, user_cipher);
 }
 
-void check_user(user_account *cipher_data, uint16_t *pword_len, 
-	user_account *login_attempt, uint16_t *found) {
+void check_user(user_account *cipher_data, uint32_t *pword_len, 
+	user_account *login_attempt, uint32_t *found) {
 
 	unsigned char decrypted[*pword_len];
 	decrypt(cipher_data->pword, *pword_len, decrypted);
