+++ user_accounts/mbside.h
@@ -5,9 +5,9 @@
 #include "userclass.h"
 #endif
 
-void create_user(user_account *user_data, uint16_t *pword_len, 
+void create_user(user_account *user_data, uint32_t *pword_len, 
 	user_account *user_cipher);
 
-void check_user(user_account *cipher_data, uint16_t *pword_len, 
-	user_account *login_attempt, uint16_t *found);
+void check_user(user_account *cipher_data, uint32_t *pword_len, 
+	user_account *login_attempt, uint32_t *found);
